Change click to remove for cookie pop-up + scrollDown input parameter
Instead of using clicking to agree to cookies so window dissapears. It now removes the element of the cookie window from the DOM. This method is more robust as the clicking was not properly working in some cases.

Added scrollDown as an input Paremter so that you can select wether you want the page to scroll down or not.

Put the set-up of the gif encoder piece of code behind the remove cookie window selector.